(improvement)(chat) support add model aliasList to dict  and update domainName to modelName
