+++ chat/knowledge/src/main/java/com/tencent/supersonic/knowledge/dictionary/builder/WordBuilderFactory.java
@@ -15,7 +15,7 @@ public class WordBuilderFactory {
     static {
         wordNatures.put(DictWordType.DIMENSION, new DimensionWordBuilder());
         wordNatures.put(DictWordType.METRIC, new MetricWordBuilder());
-        wordNatures.put(DictWordType.DOMAIN, new DomainWordBuilder());
+        wordNatures.put(DictWordType.DOMAIN, new ModelWordBuilder());
         wordNatures.put(DictWordType.ENTITY, new EntityWordBuilder());
         wordNatures.put(DictWordType.VALUE, new ValueWordBuilder());
     }
